amplify/yolo-lambda-inference/app/requirements.txt
ultralytics==8.3.16
torch
numpy
Pillow
boto3